Add support for multipackage installs
Chef 12.1.0 introduced "multipackage" installation, where the package
resource can take the name of packages to install as an array.

This change introduces a new library helper method,
multipackage_supported?, which returns true or false depending on
whether it's supported.

There isn't an easy way to duck-type whether the package provider on
an arbitrary platform supports multipackage. While we know it's
available in Chef 12.1.0+, it is better to ask for the feature,
rather than ask for the Chef version. As other package managers may have
support for multipackage added in the future, we want to ensure that
this works without having to do version comparisons.

This commit also adds test kitchen support, so we can check
12.0.3 where multipackage was NOT supported, and 12.3.0 (the current
release), where multipackage IS supported for the two platforms included
in the test matrix (apt on ubuntu, yum on centos).